+++ springsecurity/democode/src/main/java/com/jiangzh/springboot/springsecurity/democode/dao/pojo/AdminUser.java
@@ -7,7 +7,10 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
 import javax.persistence.Table;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Data;
+import lombok.NoArgsConstructor;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
@@ -20,7 +23,9 @@ import org.springframework.security.core.userdetails.UserDetails;
  * @date : 2019-06-06 13:56
  **/
 @Data
-@RequiredArgsConstructor
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
 @Entity
 @Table(name = "admin_user")
 public class AdminUser implements UserDetails {